Use udevadm settle instead of sync
Doing an explicit sync is needed only if you remove the USB drive
without unmounting it.  Since the script unmounts it anyway, and since,
removing the USB drive while the script is running will always produce a
bad image, we can remove the syncs and speed up the script.

That said, it *is* important to wait for partitions to show up after
making changes to the device. For this reason, add calls to "udevadm
settle", which does precisely this and is typically much faster than a
sync.