Use hdparm -r for checking writability
Currently, the writability check can fail if you flash an unformatted,
unmounted SD card. This is because /etc/mtab (and /proc/mounts) are used
for tracking mounted partitions. In the case of an unformatted volume,
nothing will show up in either of the files. So instead, use hdparm -r
to check for writability.