A few assorted fixes and cleanups
* Use udevadm settle instead of sync

Doing an explicit sync is needed only if you remove the USB drive
without unmounting it.  Since the script unmounts it anyway, and since,
removing the USB drive while the script is running will always produce a
bad image, we can remove the syncs and speed up the script.

That said, it *is* important to wait for partitions to show up after
making changes to the device. For this reason, add calls to "udevadm
settle", which does precisely this and is typically much faster than a
sync.

* Set -eo pipefail for safety

These provide a few safety checks against things going wrong:

-e: stop if a command returns a non-0 status
-o pipefail: chain failures through pipes so that if one program in the
pipe fails, the entire pipe is treated as failing

* Use hdparm -r for checking writability

Currently, the writability check can fail if you flash an unformatted,
unmounted SD card. This is because /etc/mtab (and /proc/mounts) are used
for tracking mounted partitions. In the case of an unformatted volume,
nothing will show up in either of the files. So instead, use hdparm -r
to check for writability.

* Typo fix

* Add Martin Kelly to AUTHORS

* Remove wait for SD card insertion

There is no quick, easy way to universally wait for and identify an SD
card on Linux. Thus in order to avoid confusing behavior, remove the
wait for SD card and instead let the user try again after inserting an
SD card.