Set date/time in cloud-init runcmd module as part on initial boot
As the Raspberry Pi's don't have a Real Time Clock (RTC), it takes
a while before `ntpd` sets the time to the current (real) time.
This creates a problem as `apt-get update` relies on the date/time
being not too far off the mark.
This commit will add a command as part of the `runcmd` section of
`cloud-init` that sets the date/time to the date/time at which the
`ssd` was flashed. Even though this will not set the date accurately,
the skew is small enough to make `apt-get update` work.