Fix date format issue
Formatting of a date as generated by the 'date' command depends
on the 'LANG' setting.

Before this fix, proper operation of the flash command depended
on the 'LANG' setting on the account from which the flash command is
run is one of a series of specific values (such as "C" or "en_US.UTF-8").

When 'cloud-init' runs the bootcmd phase, the date is set using the
'date -s' command. At that time, 'LANG' has as value "en_US.UTF-8".